Add verification test for player class
The verification in the index.test.js file just verifies if the name is inserted by the user or not. This because we need to separate the computer user from the normal user that plays the game. If the name is not inserted the game automatically takes this as a creation of a computer user, otherwise the name inserted is taken and saved in the name variable of object