Add fillBoards function to fill the squares with ships
The function prints correctly but there is a logical error with 2d arrays. In particular the function uses a unique for loop. So this loop just writes on the first row of the gameBoards because it's the only index given to the gameBoard class. A column needs to be added so that it prints correctly in all the squares