Add images for ships into javascript
The images gets added only with source that gets taken from the original image. Maybe changing the webpack order of taking files could work but it's not certain. Also a color of text of 0,0,0,0 rgba is added into the style.css file so that the textes doesn't show on screen